Move network-libp2p into network
