--- core/network-libp2p/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-description = "libp2p implementation of the ethcore network library"
-homepage = "http://parity.io"
-license = "GPL-3.0"
-name = "substrate-network-libp2p"
-version = "2.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-
-[dependencies]
-byteorder = "1.3"
-bytes = "0.4"
-fnv = "1.0"
-futures = "0.1"
-libp2p = { version = "0.9.1", default-features = false, features = ["secp256k1", "libp2p-websocket"] }
-parking_lot = "0.8.0"
-lazy_static = "1.2"
-log = "0.4"
-rand = "0.6"
-serde = { version = "1.0.70", features = ["derive"] }
-serde_json = "1.0.24"
-smallvec = "0.6"
-substrate-peerset = { path = "../peerset" }
-tokio-io = "0.1"
-tokio-timer = "0.2"
-unsigned-varint = { version = "0.2.1", features = ["codec"] }
-void = "1.0"
-zeroize = "0.6.0"
-
-slog = { version = "^2", features = ["nested-values"] }
-slog_derive = "0.1.1"
-erased-serde = "0.3.9"
-
-[dev-dependencies]
-tempdir = "0.3"
-tokio = "0.1"
