+++ core/network/src/debug_info.rs
